+++ life_star.js
@@ -9,10 +9,9 @@ var express = require('express'),
     auth = require('./lib/auth'),
     WorkspaceHandler = require('./lib/workspace').WorkspaceHandler,
     SubserverHandler = require('./lib/subservers').SubserverHandler,
+    ManifestHandler = require('./lib/manifest').ManifestHandler,
     spawn = require('child_process').spawn,
-    exec = require('child_process').exec,
-    fs = require('fs'),
-    path = require('path');
+    fs = require('fs');
 
 module.exports = function serverSetup(config) {
 
@@ -124,36 +123,12 @@ module.exports = function serverSetup(config) {
   // -=-=-=-=-=-=-=-
   new SubserverHandler({baseURL: '/nodejs/', additionalSubservers: config.subservers}).registerWith(app);
 
-  // -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
-  // manifest related -- manifest files are used by web browsers to cache
-  // files
-  // -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
-  var manifestFileName = "lively.scriptscache";
-  if (config.useManifestCaching) {
+  // -=-=-=-=-=-=-=-=-=-=-
+  // manifest file related
+  // -=-=-=-=-=-=-=-=-=-=-
+  var manifestHandler = new ManifestHandler(config);
+  manifestHandler.registerWith(app);
 
-    var rootDir = config.fsNode;
-    function findJSFilesForManifest(thenDo) {
-      exec('find . -type f -iname "*.js" -print', {cwd: rootDir}, function(code, out, err) {
-        thenDo(code, out);
-      });
-    }
-
-    app.get('/' + manifestFileName, function(req, res) {
-      // res.type('text/cache-manifest');
-      findJSFilesForManifest(function(err, filesString) {
-        if (err) {
-          res.status(500).send('');
-          return;
-        }
-        filesString = filesString.replace(/\.\//g, '/');
-        res.set({
-          'Content-Type': 'text/cache-manifest',
-          'Cache-Control': 'no-cache, private'
-        });
-        res.send("CACHE MANIFEST\n# version\n\n" + filesString);
-      });
-    });
-  }
   // -=-=-=-=-=-
   // set up DAV
   // -=-=-=-=-=-
@@ -166,36 +141,11 @@ module.exports = function serverSetup(config) {
   if (!srv.baseUri) srv.baseUri = '/';
   if (!srv.getBaseUri) srv.getBaseUri = function() { return this.baseUri };
 
-  function addManifestRef(req, res) {
-    if (!(/\.html$/.test(req.url)) || req.method.toLowerCase() !== 'get') return;
-    var interceptedHeaders, interceptedHeadersCode,
-        writeFunc = res.write,
-        writeHeadFunc = res.writeHead;
-    res.writeHead = function(code, headers) {
-      interceptedHeaders = headers;
-      interceptedHeadersCode = code;
-      if (code >= 400) {
-        writeHeadFunc.call(this, code, headers);
-      }
-    }
-    res.write = function(data) {
-      if (interceptedHeadersCode >= 400) {
-        return writeFunc.call(this, data);
-      }
-      var s = data.toString();
-      s = s.replace(/<html>/, '<html manifest="' + manifestFileName + '">');
-      interceptedHeaders['content-length'] = s.length;
-      writeHeadFunc.call(this, interceptedHeadersCode, interceptedHeaders);
-      return writeFunc.call(this, s);
-    }
-  }
-
-
   function fileHandler(req, res) {
     if (req.url.match(/\?\d+/)) {
       req.url = req.url.replace(/\?.*/, ''); // only the bare file name
     }
-    if (config.useManifestCaching) addManifestRef(req, res);
+    manifestHandler.addManifestRef(req, res);
     new DavHandler(srv, req, res);
   };
